refactor BrokeragePresenter and BrokerageController
